v4.7
Re-added Encourage/Discourage buttons outside of Advanced Mode, and added New Session button to Advanced Mode. Both Encourage and Discourage buttons will now start a new session after doing their thing, while the New Session button will only start a new session (doesn't discourage anymore). Also fixed the settings (Config.ini file) getting reset to default when erasing the AI's memory.

+++ app/src/main/java/com/oblivionburn/nlp/Util.java
@@ -359,7 +359,11 @@ class Util
                 EraseMemory(child);
             }
         }
-        fileOrDirectory.delete();
+
+        if (!fileOrDirectory.getName().contains("Config"))
+        {
+            fileOrDirectory.delete();
+        }
     }
 
     static void ToggleAdvanced(MenuItem item)
